fetch userlist and bondlist
